Moving common QA class helpers to helper file
